fix(fmath): vendor architecture-deterministic float32 transcendentals (#79) (#106)
* fix(fmath): vendor architecture-deterministic float32 transcendentals (#79)

The encoder's float32 transcendentals wrapped package math, whose
Exp/Exp2/Log/Log2/Pow/Atan/Sin/Cos dispatch to per-architecture assembly or to
a portable Go path that gc contracts into FMADDD on arm64. The same input
therefore produced last-ulp-different float64 results per architecture, and a
fused a*b+c also disagrees with the -ffp-contract=off C reference. Those
sub-ULP differences round away in the float32 pipeline today, so encoder output
is unchanged, but they are a latent reproducibility hazard: a different amd64
CPU, a GOAMD64=v3 build, or a stdlib change could tip a rounded float32 and
perturb the bitstream.

Vendor barriered pure-Go copies of the Go 1.27.0 stdlib math (exp, exp2,
expmulti, log, log2, pow, atan, sin, cos, trigReduce) into internal/fmath, with
an explicit float64() rounding barrier on every internal a*b+c and //go:noinline
on the entry points so an inlined product cannot fuse with a caller's add.
Rewire the fmath wrappers and the encoder-only math.Sin/Cos sites in
internal/window, internal/mdct and internal/dsp/lpc.go, and barrier the float64
multiply-adds in the MDCT twiddle path, the TNS autocorrelation and Levinson
recursion, and BesselI0's series so the whole determinism-critical float64
surface is FMA-free.

The barriered copies are bit-identical to the amd64 stdlib for the fusion-only
functions (Sin/Cos/Atan/Exp2), and Exp/Log/Log2/fractional Pow change both
arches to one deterministic value; all eleven vendored-function hashes are
identical on arm64 and amd64. Encoder output stays byte-identical to before on
the corpus (the change rounds away at float32) and remains
architecture-deterministic.

Add TestNoFloat64FMAInDeterministicMath (internal/fmath, internal/mdct and
internal/dsp compile with no float64 FMA on arm64), pin the vendored functions'
architecture-independent output hashes, and pin the four generated window
tables. Include the Go BSD-3-Clause license as LICENSE.golang and document the
derivation in PROVENANCE.md.

* fix(psy): barrier the ATH curve and add a transcendental accuracy test (#79)

Address review feedback on the architecture-determinism guard's coverage.

internal/psy ath computes the absolute-threshold-of-hearing curve in float64 and
casts it to the float32 psy threshold, but its four multiply-adds were left
unbarriered, so gc fused them into FMADDD on arm64: the same latent architecture
split the vendored transcendentals eliminate, in one package the float64 FMA
guard did not cover. Barrier each product (value-neutral on amd64, which does not
fuse at GOAMD64=v1; encoder output stays byte-identical on the corpus on both
arches) and extend TestNoFloat64FMAInDeterministicMath to internal/psy, so the
guarded surface (fmath, mdct, dsp, psy) is now the encoder's complete
determinism-critical float64 set. internal/coder and internal/enc are already
float64-FMA-free.

Add TestPortableAccuracy, which compares every vendored transcendental against
the standard library over an input sweep within a ULP tolerance.
TestPortableDeterminismGolden only hashes the current implementation, so a
coefficient transcription error would hash stably; the accuracy test catches it,
since a wrong constant moves the result far more than the intended sub-ULP
barrier and FMA difference. Verified by mutation: flipping an exp coefficient
turns it red.

+++ internal/psy/psy.go
@@ -207,13 +207,19 @@ const athAdd = 4
 // borrowed from LAME. Mirrors aacpsy.c:ath @ d09d5afc3a: the frequency
 // scaling is float32, the curve itself is computed in float64 like the C
 // (double constants promote the expression).
+//
+// Each term's product carries a float64() barrier so gc does not fuse it into
+// an FMADDD on arm64: the C reference is -ffp-contract=off, and the float64
+// curve is cast to float32 into the psy threshold, so a fused sub-ulp drift is
+// a latent architecture split (issue #79). The value is unchanged on amd64,
+// which does not fuse at GOAMD64=v1.
 func ath(f, add float32) float32 {
 	f /= 1000.0
 	fd := float64(f)
-	v := 3.64*fmath.Pow(fd, -0.8) -
-		6.8*fmath.Exp(-0.6*(fd-3.4)*(fd-3.4)) +
-		6.0*fmath.Exp(-0.15*(fd-8.7)*(fd-8.7)) +
-		(0.6+0.04*float64(add))*0.001*fd*fd*fd*fd
+	v := float64(3.64*fmath.Pow(fd, -0.8)) -
+		float64(6.8*fmath.Exp(-0.6*(fd-3.4)*(fd-3.4))) +
+		float64(6.0*fmath.Exp(-0.15*(fd-8.7)*(fd-8.7))) +
+		float64((0.6+float64(0.04*float64(add)))*0.001*fd*fd*fd*fd)
 	return float32(v)
 }
